requirements.txt
langchain
pyodbc
python-dotenv
openai
langchain_openai
SQLAlchemy
langchain-community
langchain-experimental
flask